refactor: rename "GridLayout" folder to "Grid"
