+++ src/Components/Form/Form.stories.tsx
@@ -25,7 +25,7 @@ import {FormElement} from 'src/Components/Form/FormElement'
 import {FormHelpText} from 'src/Components/Form/FormHelpText'
 import {FormElementError} from 'src/Components/Form/FormElementError'
 import {FormValidationElement} from 'src/Components/Form/FormValidationElement'
-import {Grid} from 'src/Components/GridLayout/Grid'
+import {Grid} from 'src/Components/Grid/Grid'
 import {Button} from 'src/Components/Button/Composed/Button'
 import {Input} from 'src/Components/Inputs/Input'
 import {FlexBox} from 'src/Components/FlexBox/FlexBox'
