+++ src/Components/Grid/Grid.stories.tsx
@@ -9,15 +9,15 @@ import {mapEnumKeys} from 'src/Utils/storybook'
 import {jsxDecorator} from 'storybook-addon-jsx'
 
 // Components
-import {Grid} from 'src/Components/GridLayout/Grid'
-import {GridRow} from 'src/Components/GridLayout/GridRow'
-import {GridColumn} from 'src/Components/GridLayout/GridColumn'
+import {Grid} from 'src/Components/Grid/Grid'
+import {GridRow} from 'src/Components/Grid/GridRow'
+import {GridColumn} from 'src/Components/Grid/GridColumn'
 
 // Types
 import {Columns} from 'src/Types'
 
 // Notes
-import GridReadme from 'src/Components/GridLayout/Grid.md'
+import GridReadme from 'src/Components/Grid/Grid.md'
 
 const gridStories = storiesOf('Layout|12 Column Grid', module)
   .addDecorator(withKnobs)
