+++ src/index.ts
@@ -39,9 +39,9 @@ export * from './Components/Form/FormFooter'
 export * from './Components/Form/FormHelpText'
 export * from './Components/Form/FormLabel'
 export * from './Components/Form/FormValidationElement'
-export * from './Components/GridLayout/Grid'
-export * from './Components/GridLayout/GridColumn'
-export * from './Components/GridLayout/GridRow'
+export * from './Components/Grid/Grid'
+export * from './Components/Grid/GridColumn'
+export * from './Components/Grid/GridRow'
 export * from './Components/Icon/Icon'
 export * from './Components/IndexList/IndexList'
 export * from './Components/IndexList/IndexListBody'
